pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.155.0",
]
